+++ darwin/dataset/__init__.py
@@ -1,1 +1,2 @@
 from darwin.dataset.remote_dataset import RemoteDataset  # noqa
+from darwin.dataset.local_dataset import LocalDataset  # noqa
